Add animation to loader and empty screen states Move strings to strings.xml (I miss you phrase plugin)
