+++ app/src/main/java/com/toadfrogson/clownhub/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import com.toadfrogson.clownhub.app.screen.MainScreen
+import com.toadfrogson.clownhub.ui.screen.MainScreen
 import com.toadfrogson.clownhub.ui.theme.ClownHubTheme
 
 class MainActivity : ComponentActivity() {
